# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AppUserRole = typeof AppUserRole[keyof typeof AppUserRole];


export const AppUserRole = {
  student: 'student',
  instructor: 'instructor',
  admin: 'admin',
} as const;

export interface AppUser {
  id: string;
  name: string;
  email: string;
  role: AppUserRole;
  joinDate: string;
  isActive: boolean;
}

export type CreateUserRole = typeof CreateUserRole[keyof typeof CreateUserRole];


export const CreateUserRole = {
  student: 'student',
  instructor: 'instructor',
  admin: 'admin',
} as const;

export interface CreateUser {
  id: string;
  name: string;
  email: string;
  passwordHash?: string;
  role: CreateUserRole;
  joinDate: string;
  isActive: boolean;
}

export type LessonType = typeof LessonType[keyof typeof LessonType];


export const LessonType = {
  video: 'video',
  reading: 'reading',
  quiz: 'quiz',
} as const;

export interface Lesson {
  id: string;
  courseId: string;
  title: string;
  duration: string;
  type: LessonType;
  order: number;
}

export type CreateLessonType = typeof CreateLessonType[keyof typeof CreateLessonType];


export const CreateLessonType = {
  video: 'video',
  reading: 'reading',
  quiz: 'quiz',
} as const;

export interface CreateLesson {
  id: string;
  title: string;
  duration: string;
  type: CreateLessonType;
  order: number;
}

export interface Course {
  id: string;
  title: string;
  description: string;
  instructorId: string;
  instructorName: string;
  category: string;
  level: string;
  duration: string;
  color: string;
  rating: number;
  isPublished: boolean;
  createdAt: string;
  lessons: Lesson[];
  enrolledStudents: string[];
}

export interface CreateCourse {
  id: string;
  title: string;
  description: string;
  instructorId: string;
  instructorName: string;
  category: string;
  level: string;
  duration: string;
  color: string;
  rating?: number;
  isPublished?: boolean;
  createdAt: string;
}

export interface UpdateCourse {
  title?: string;
  description?: string;
  category?: string;
  level?: string;
  duration?: string;
  color?: string;
  rating?: number;
  isPublished?: boolean;
}

export interface EnrollRequest {
  studentId: string;
}

export interface Assignment {
  id: string;
  courseId: string;
  courseName: string;
  title: string;
  description: string;
  dueDate: string;
  maxScore: number;
}

export interface CreateAssignment {
  id: string;
  courseId: string;
  courseName: string;
  title: string;
  description?: string;
  dueDate: string;
  maxScore: number;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  submitted: 'submitted',
  graded: 'graded',
} as const;

export interface Submission {
  id: string;
  assignmentId: string;
  studentId: string;
  studentName: string;
  answer: string;
  score?: number;
  status: SubmissionStatus;
  submittedAt: string;
}

export interface SubmitRequest {
  id: string;
  studentId: string;
  studentName: string;
  answer: string;
  submittedAt: string;
}

export interface GradeRequest {
  score: number;
}

export interface ProgressResult {
  userId: string;
  courseId: string;
  completedLessons: string[];
  percentage: number;
}

export interface MarkProgressRequest {
  id: string;
  userId: string;
  courseId: string;
  lessonId: string;
  completedAt: string;
}

export type GetSubmissionsParams = {
assignmentId?: string;
studentId?: string;
};