# lib/api-zod/src/generated/types/appUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */
import type { AppUserRole } from './appUserRole';

export interface AppUser {
  id: string;
  name: string;
  email: string;
  role: AppUserRole;
  joinDate: string;
  isActive: boolean;
}