# lib/api-zod/src/generated/types/createCourse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCourse {
  id: string;
  title: string;
  description: string;
  instructorId: string;
  instructorName: string;
  category: string;
  level: string;
  duration: string;
  color: string;
  rating?: number;
  isPublished?: boolean;
  createdAt: string;
}