# lib/api-zod/src/generated/types/markProgressRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */

export interface MarkProgressRequest {
  id: string;
  userId: string;
  courseId: string;
  lessonId: string;
  completedAt: string;
}