# lib/api-zod/src/generated/types/progressResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressResult {
  userId: string;
  courseId: string;
  completedLessons: string[];
  percentage: number;
}