# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionStatus } from './submissionStatus';

export interface Submission {
  id: string;
  assignmentId: string;
  studentId: string;
  studentName: string;
  answer: string;
  score?: number;
  status: SubmissionStatus;
  submittedAt: string;
}