# lib/api-zod/src/generated/types/submitRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnHub Online Learning System API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitRequest {
  id: string;
  studentId: string;
  studentName: string;
  answer: string;
  submittedAt: string;
}